Service zum durchführen der Quiz erstellt und interface Clock für kontrollierbare Zeit in Test

+++ src/main/java/de/propra/exam/domain/service/QuizRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface QuizRepository {
     List<Quiz> findAll();
     void save(Quiz quiz);
-    Optional<Quiz> findById(Integer id);
+    Optional<Quiz> findById(Long id);
 }
